pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "tracerag"
version = "0.3.0"
description = "TraceRAG milestone 3: hybrid retrieval"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "typer>=0.12.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "pypdf>=4.2.0",
  "beautifulsoup4>=4.12.0",
  "httpx>=0.27.0",
  "faiss-cpu>=1.8.0",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0"
]

[project.scripts]
tracerag = "tracerag.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["tracerag*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
